Settings: removed usage of jsonParam CSS class
The jsonParam marker class is used in the cloud API to mark field that must
be processed as JSON data. This seems only to affect so far the way properties
modification are detected (in order to display the "unsaved changes" yellow
bar).

Problem: the function detecting field changes is not able to handle unitialized
fields in the "initial" cloud profile configuration, this will be a problem for
existing cloud profiles when no cloud profile assocation have been made yet.