Gracelul shutdown of idle TeamCity agents
The container stop-timeout can now be configured in the agent image settings.
This timeout will be applied when stopping individual containers either when
the agent max idle time is reached, or if the agent was explicitly shut down
in the cloud management interface.

When no timeout is specified in the image configuration, the daemon default
timeout will apply.

The timeout used to discard containers when the cloud profile is disposed
(either upon reconfiguration or when the TC instance is being shut down) is
always forced to 0 seconds in order to emit all stop requests ASAP  before a
possible halt of the JVM.

Closes #9