Fix volume binding with colon in file paths
The volume bindings definition as specified in the Docker API cannot be
reliably parsed back if the host or container path is containing a colon
character, which is also used as field delimiter in the binding
definition.

To handle this, the volume binding specifications are now persisted in
the Editor section of the cloud profile configuration as well, using
specific JSON fields to seperate the host and container paths.

Cloud profile settings have been upgraded to version 2.