Support for multiple engine API version
Docker engine API version support is now defined with a minimal version
in addition to the default target version. When checking for
connectivity with the Docker daemon, a warning will be issued only if
the effective API version (after negotiation) is outside of this range.

API version check is now performed on the client side, since it will
later be performed on individual container configuration fields as
well during the edition of the container settings.

For convenience, the last known engine API version is saved as a new
configuration property, so the user don't have to fetch the engine API
version each time the plugin settings are being edited.

The whole daemon "version" JSON structure is actually also persisted in
the plugin configuration. This is used to display helpful information
to the user and include additional meta-data such as OS name, that will
be useful to build further validation logic.