Build: upgraded TC API to version 2017.1
There is no change in the APIs we are using. The server JAR is however
not anymore a dependency of the server API and must be explicitely
declared using another repository artifacts.

Added comment regarding pending feature request that are required to
remove our dependence on the server JAR (not part of the public API).