Cloud client: fix potential dead-lock
The cloud cloud client is currently using the agent manager to remove
obsolete agent instances. To do so, the agent manager will require to
lock the RunningInstanceTracker class of the TC server.

Concurrently, the instance tracker may in turn also try to access the
cloud client, creating a possibility of dead-locks.

Fix #18